Move planning components to new plans folder
